// frontend/src/shared/api/generated/model/tradingSignalResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * StockQuest API
 * Stock Trading Simulation Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { MarketConditionResponse } from "./marketConditionResponse";
import type { TradingSignalResponseSignalType } from "./tradingSignalResponseSignalType";
import type { TradingSignalResponseStatus } from "./tradingSignalResponseStatus";
import type { TradingSignalResponseStrength } from "./tradingSignalResponseStrength";
import type { SignalReasonResponse } from "./signalReasonResponse";

export interface TradingSignalResponse {
  /** 신뢰도 */
  confidence?: string;
  /** 예상 수익률 */
  expectedReturn?: string;
  /** 예상 위험도 */
  expectedRisk?: string;
  /** 시그널 만료 시간 */
  expiresAt?: string;
  /** 시그널 생성 시간 */
  generatedAt?: string;
  marketCondition?: MarketConditionResponse;
  /** 시그널 ID */
  signalId?: string;
  /** 시그널 타입 */
  signalType?: TradingSignalResponseSignalType;
  /** 시그널 상태 */
  status?: TradingSignalResponseStatus;
  /** 손절 가격 */
  stopLossPrice?: string;
  /** 시그널 강도 */
  strength?: TradingSignalResponseStrength;
  /** 주식 심볼 */
  symbol?: string;
  /** 목표 가격 */
  targetPrice?: string;
  /** 투자 기간 (일) */
  timeHorizon?: number;
  /** 주요 근거 */
  topReasons?: SignalReasonResponse[];
}